Get album name using Azuracast's NowPlaying API
This method is... more than a little hacky, it pulls the /api/nowplaying/
endpoint, finds the station data based on name, and then pulls album
data from there. But for some reason, the XMLHttpRequest is always
sending an HTTP request, even when manually specifying a full URL,
'https://' included. So, you'll get one mixed content warning per
second, but it will somehow get at least one that goes through and
pulls the album name as expected.

So, it works, but I have no idea why, and nothing I can control seems
to be able to fix it.